--- a/monix-execution/js/src/main/scala/monix/execution/internal/collection/queues/ConcurrentQueueBuilders.scala
+++ b/monix-execution/js/src/main/scala/monix/execution/internal/collection/queues/ConcurrentQueueBuilders.scala
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2014-2018 by The Monix Project Developers.
+ * See the project homepage at: https://monix.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package monix.execution.internal.collection.queues
+
+import monix.execution.ChannelType
+import monix.execution.internal.collection.{ArrayQueue, ConcurrentQueue}
+
+private[internal] trait ConcurrentQueueBuilders {
+  /**
+    * Builds a `ConcurrentQueue` reference.
+    */
+  def apply[A](capacity: Int, channelType: ChannelType): ConcurrentQueue[A] =
+    ArrayQueue.bounded[A](capacity)
+}